Add a -capability command-line option (#1651)
This provides a stand-alone option distinct from `-profile` that can be used to add capabilities to a target. A test has been added to confirm that `-profile X -capability Y` works the same as `-profile X+Y`.

The intention is that this option could be used in applications that use the API to set up their target but then use the options-parsing logic to handle things like capabilities.

Note: that latter bit has not been confirmed, so it is possible that this approach does not actually suffice for hybrid API + options usage. That will need to be confirmed in follow-up work.