[gfx] Shader-object driven shader compilation.
Adds the following responsibilities to `gfx` layer:
- Shader/pipeline specialization based on the type of bound shader-objects
- Caching of specialized shader kernels
Note: none of these new `gfx` capabilities are used and tested by any
part of the code base, and the implementation and interface involving
these new capabilities are likely to be changed in future commits.
The implementation only handles the simplest case of shader specialization:
via a sub-object defined by an interface-typed field.
A `ParameterBlock<IInterface>` or `StructuredBuffer<IInterface>` typed
field is not implemented yet. As a result, one of the dynamic-dispatch
tests has to be disabled for now since it uses `ConstantBuffer<IInterface>`.

Refactors to the `gfx` layer:
- Remove `gfxGetCreateFunc` and just expose `gfxCreateRenderer`.
- Remove `IRenderer::initialization` method and performs initialization
  directly at renderer creation time.
- Move many logic in `renderer-graphics-common.h/cpp` to `renderer-shared.h/cpp`,
  where things in `renderer-graphics-common` are shared by all graphics
  API implementations, and things in `renderer-shared` are shared by all
  renderer implementations, including CUDA and CPU. This increases code
  reuse and allow most shader compilation and specialization logic to be
  shared among all renderers.
- `IRenderer::setPipelineState` no longer takes a `PipelineType` parameter.

General changes:
- Change from C style type cast to `static_cast<RefObject*>` in
  `RefPtr<T>::releaseReference` to capture all `RefPtr` mis-uses.