Streamline shader object creation (#1717)
This change kind of rolls together two different simplifications:

1. The `createShaderObject()` shouldn't really need to take an `IShaderObjectLayout` because it could just take the `slang::TypeLayoutReflection` instead and create the shader-object layout behind the scenes.

2. For that matter, it needn't take a `slang::TypeLayoutReflection` either, becaues it could just take a `slang::TypeReflection` and query the layout of that type behind the scenes.

The combination of these two changes means:

* `IShaderObjectLayout` is gone from the public API, as is `createShaderObjectLayout()`

* `createShaderObject()` directly takes a `slang::TypeReflection` and allocates a shader object of that type

The result is simpler and more streamlined application code.

Note that under the hood the implementation still has shader-object layouts, using the `ShaderObjectLayoutBase` class. A few locations had to change to use `RefPtr`s instead of `ComPtr`s now that the class is no longer a public COM-lite API type.

The hope is that this change makes it easier to allocate/cache layouts for things like specialized types "under the hood," as is needed to implement parameter setting for static specialization.